refactor: Home Folder Structure
