+++ src/components/HomeSection/index.tsx
@@ -1,0 +1,4 @@
+export { Home } from './Home';
+export { TimeLine } from './TimeLine';
+export { TimeLineTransition } from './TimeLineTransition';
+export { TypingAnimation } from './TypingAnimation';
